Listener: use class UniqueSocketDescriptor

+++ src/Listener.hxx
@@ -6,10 +6,12 @@
 
 #pragma once
 
+#include <cstdint>
+
 struct Instance;
 
 void
-listener_init(Instance *instance, unsigned port);
+listener_init(Instance *instance, uint16_t port);
 
 void
 listener_deinit(Instance *instance);
